handle failure reading non-200 response body
If the response body  of a non-200 failed to deserialize, the type would change
from `Future[(Failure[HttpStatusException], Promise])` to
`Future[Failure[IOException]]`.

This commit restores the balance. A response parsing exception is now
propagated inside the HttpStatusException, so that neither the non-200 status
code nor the `IOException` are lost.